SE2R-40: implement recipe add, initial commit

+++ backend/src/main/java/com/recipes/backend/rest/RecipesController.java
@@ -1,11 +1,17 @@
 package com.recipes.backend.rest;
 
+import static com.recipes.backend.mapper.RecipeMapper.mapToRecipe;
+import static com.recipes.backend.utils.LogWriter.logHeaders;
+
 import com.recipes.backend.bizz.recipe.RecipeService;
+import com.recipes.backend.bizz.security.SecurityService;
+import com.recipes.backend.exception.domain.RecipeEmptyException;
 import com.recipes.backend.mapper.RecipeMapper;
 import com.recipes.backend.rest.domain.RecipeAllRest;
 import com.recipes.backend.rest.domain.RecipeRest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -18,17 +24,23 @@ import java.util.stream.Collectors;
 public class RecipesController {
 
     private final RecipeService recipeService;
+    private final SecurityService securityService;
 
     @Autowired
-    public RecipesController(final RecipeService recipeService) {
+    public RecipesController(final RecipeService recipeService, final SecurityService securityService) {
         this.recipeService = recipeService;
+        this.securityService = securityService;
     }
 
     @GetMapping
     public ResponseEntity<RecipeAllRest> getAllRecipes(@RequestHeader HttpHeaders headers,
                                                        @RequestParam(name = "page") int page,
                                                        @RequestParam(name = "limit") int limit) {
-        //TODO check headers
+        logHeaders(headers);
+
+        if (!securityService.isAuthenticated(headers)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
 
         final Set<RecipeRest> retrievedRecipes =
                 recipeService.getAllRecipes(page, limit).stream()
@@ -41,10 +53,29 @@ public class RecipesController {
         return ResponseEntity.ok(new RecipeAllRest(totalRecipes, retrievedRecipes));
     }
 
+    @PostMapping
+    public ResponseEntity<Object> addRecipe(@RequestHeader HttpHeaders headers, @RequestBody RecipeRest recipeRest) {
+        logHeaders(headers);
+
+        if (!securityService.isAuthenticated(headers)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        final var recipeToAdd = mapToRecipe(recipeRest).orElseThrow(RecipeEmptyException::new);
+        recipeService.addRecipe(recipeToAdd);
+
+        return ResponseEntity.ok().build();
+    }
+
     @DeleteMapping("/{id}")
     public ResponseEntity<String> deleteRecipe(@RequestHeader HttpHeaders headers,
                                                @PathVariable(name = "id") Long recipeId) {
-        //TODO add header validation
+        logHeaders(headers);
+
+        if (!securityService.isAuthenticated(headers)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         return recipeService.deleteRecipe(recipeId) ? ResponseEntity.ok(recipeId.toString()) : ResponseEntity.badRequest().body("Bad request!");
     }
 }
